Rearrange file structure and routes to accommodate for a separate all programs for students and admin
